% Distance matrix Rcpp
\name{distMatRcpp}
\alias{distMatRcpp}
\title{This function computes the pair wise distances of tips according to the branching order using Rcpp.}
\usage{
distMatRcpp(len, len2, nHaplos, mD)
}
\arguments{
\item{len}{Length of tip set 1.}

\item{len2}{Length of tip set 2.}

\item{nHaplos}{Number of haplotypes.}

\item{mD}{Distance matrix}
}

\description{
This function computes pair-wise distances.
}

\keyword{internal}