% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data.R
\docType{data}
\name{ex_hapMat_data}
\alias{ex_hapMat_data}
\title{Example dataset}
\format{A list of 200 haplotypes with the physical positions of each SNV.
\describe{
\item{hapmat}{A matrix of 0's and 1's, with rows representing haplotypes and columns representing SNVs.}
\item{snvNames}{A vector of names of SNVs for the columns of \code{hapmat}.}
\item{hapNames}{A vector of names of haplotypes for the rows of \code{hapmat}.}
\item{posns}{ A numeric vector specifying the genomic positions (e.g. in base pairs) of SNVs in the
 columns of \code{hapmat}.}
}}
\usage{
data(ex_hapMat_data)
}
\description{
A \code{hapMat} data object containing 200 sequences (haplotypes) with 2747 SNVs.
}
\keyword{datasets}
