% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_hapMat_utility_function.R
\name{extractHaplos}
\alias{extractHaplos}
\title{Extract haplotypes and return it as a list from genotype matrix by vcf file.}
\usage{
extractHaplos(vcfGeno, snvPosns)
}
\arguments{
\item{vcfGeno}{genotype matrix extract from vcf file.}

\item{snvPosns}{SNV positions in base paris.}
}
\value{
a list of haplotype data with positions
}
\description{
Extract haplotypes and return it as a list from genotype matrix by vcf file.
}
\keyword{internal}
