% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_SNV_utility_functions.R
\name{getnSNVs}
\alias{getnSNVs}
\title{Number of SNVs}
\usage{
getnSNVs(cur)
}
\arguments{
\item{cur}{A numeric vector of two elements indexing the first and last SNV.}
}
\value{
A numeric value specifying the number of SNVs returned to \code{\link{findSNVs}}
}
\description{
This function gives the number of SNVs contained within a given pair of indices of SNVs, inclusively.
}
\examples{
SNV_index = c(2,5)
getnSNVs(SNV_index) # return 4. i.e. number of SNVs from SNV2 to SNV5, inclusively is 4.


}
\seealso{
\code{\link{findSNVs}}
}
\keyword{internal}
