% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dend_utility_functions.R
\name{makeNewickRec}
\alias{makeNewickRec}
\title{Build the character string of nodes and haplotypes in Newick format}
\usage{
makeNewickRec(node, sep)
}
\arguments{
\item{node}{Tree(dend) node}

\item{sep}{A separator for haplotype names in the same clade. See the arguments in
\code{\link{reconstructPP}}.}
}
\description{
Build the character string of nodes and haplotypes in Newick format
}
\keyword{internal}
