% From Rcpp

\name{mantelStat}
\alias{mantelStat}
\title{Compute mantel statistics of two distance matirces}
\usage{
mantelStat(Dx, Dy)
}
\arguments{
\item{Dx}{A numeric matrix of pairwise distances.}
\item{Dy}{A second numeric matrix of pairwise distances.}
}

\description{This function computes the mantel statistics of two distance matrices. 
}
\references{
Mantel, N. (1967). The detection of disease clustering and a generalized
regression approach. Cancer Research, 27, 209-220.
}

\keyword{internal}

