% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dend_utility_functions.R
\name{noVariation}
\alias{noVariation}
\title{Check the variation in a SNV}
\usage{
noVariation(snv)
}
\arguments{
\item{snv}{A SNV to check for variation among haplotypes.}
}
\value{
Logical:TRUE, if there is no variation in the SNV; FALSE otherwise.
}
\description{
This function checks the variation in a specified SNV and is applied during the recursive partitioning.
}
\seealso{
\code{\link{makePartition}}
}
\keyword{internal}
