\name{perfectphyloR-package}
\alias{perfectphyloR-package}
\alias{perfectphyloR}
\docType{package}
\title{ Reconstruct perfect phylogenies from DNA sequence data

}
\description{ Functions to reconstruct perfect phylogeny underlying a sample of DNA sequences, at a focal single-nucleotide variant (SNV) and to depict and test association in a genomic region based on the reconstructed partitions.

}
\author{
Charith Karunarathna and Jinko Graham
}


