% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_tests_utility_functions.R
\name{performTest}
\alias{performTest}
\title{Perform the user-provided association test}
\usage{
performTest(Dx, Dy, testname, nperm)
}
\arguments{
\item{Dx}{A numeric matrix of pairwise distances.}

\item{Dy}{A second numeric matrix of pairwise distances.}

\item{testname}{Association methods. Use "HHG" for HHG test, "dCor" for dCOr test, "Mantel" for 
mantel test, "RV" for RV test and "RI" for Rand index test.}

\item{nperm}{Numer of permutations.}
}
\description{
This function directs to perform the user-given method.
}
\keyword{internal}
