\docType{data}
\name{phenoDist}
\alias{phenoDist}
\title{Phenotypic distances }
\format{An object of class \code{matrix}.}
\usage{
data(phenoDist)
}
\description{
This is the pairwise phenotypic distances described in Karunarathna and Graham (2018).
}
\references{
Karunarathna, C. B., & Graham, J. (2018). Using gene genealogies to localize rare variants associated with complex traits in diploid populations. Human heredity, 83(1), 30-39.
}
\keyword{datasets}
