% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_index_test.R
\name{RandIndex}
\alias{RandIndex}
\title{Rand Index}
\usage{
RandIndex(group1, group2)
}
\arguments{
\item{group1}{First group.}

\item{group2}{Second group.}
}
\value{
A numeric value between 0 and 1.
}
\description{
This function calculates the Rand index, a  measure of similarity between two
partitions and returns a value between 0 and 1 reflecting the
agreement of the partitions.
}
\examples{

\dontshow{

g1 <- sample(1:2, size=10, replace=TRUE)
g2 <- sample(1:3, size=10, replace=TRUE)

RandIndex(g1,g2)


}

}
\keyword{internal}
