% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_index_test.R
\name{splitTips}
\alias{splitTips}
\title{Separate haplotype names for haplotypes that can not be distingushed in the window around the focal point}
\usage{
splitTips(group)
}
\arguments{
\item{group}{A numeric vector that represent the group indices returned by \code{cutree()}.}
}
\value{
A vector of haplotypes.
}
\description{
This function separates haplotypes names for haplotypes, if they can not be distinguished in the window
around the focal point.
}
