% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel_test.R
\name{mantelStat}
\alias{mantelStat}
\title{This function performs Mantel test for correlation between two matrices.}
\usage{
mantelStat(Dx, Dy)
}
\arguments{
\item{Dx}{A numeric matrix of pairwise distances.}

\item{Dy}{A second numeric matrix of pairwise distances.}
}
\value{
A matrix of numeric values.
}
\description{
This function performs Mantel test for correlation between two matrices.
}
\keyword{internal}
