% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_index_test.R
\name{RandIndex}
\alias{RandIndex}
\title{Rand Index}
\usage{
RandIndex(group1, group2)
}
\arguments{
\item{dend1}{Partition 1}

\item{dend2}{Partition 2}

\item{k}{The number of clusters that the dendrogram should be cut into.}
}
\value{
A numeric value between 0 and 1.
}
\description{
This function calculates the Rand index, a  measure of similarity between two
partitions and returns a value between 0 and 1 reflecting the
agreement of the partitions.
}
\keyword{internal}
